+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This projec
 ## Development
 
 - Add .gitattributes, .editorconfig files
+- Fix route function to allow class name parameter
 
 ## v1.1.1